+++ tests/testthat/test_as_task_unsupervised.R
@@ -1,4 +1,4 @@
-test_that("as_task works on stars objects", {
+test_that("as_task_unsupervised works on stars objects", {
   skip_if_not_installed("stars")
   requireNamespace("stars", quietly = TRUE)
 
@@ -8,19 +8,19 @@ test_that("as_task works on stars objects", {
   dimension = 10)
   stack = invoke(stars::st_as_stars, .x = stack, .opts = allow_partial_matching)
 
-  expect_class(as_task_unsupervised(stack), "Task")
+  expect_class(as_task_unsupervised(stack), "TaskUnsupervised")
 })
 
-test_that("as_task works on SpatRaster objects", {
+test_that("as_task_unsupervised works on SpatRaster objects", {
   stack = generate_stack(list(
     numeric_layer("x_1"),
     numeric_layer("y")),
   dimension = 10)
 
-  expect_class(as_task_unsupervised(stack), "Task")
+  expect_class(as_task_unsupervised(stack), "TaskUnsupervised")
 })
 
-test_that("as_task works on RasterBrick objects", {
+test_that("as_task_unsupervised works on RasterBrick objects", {
   skip_if_not_installed("raster")
   requireNamespace("raster", quietly = TRUE)
 
@@ -30,10 +30,10 @@ test_that("as_task works on RasterBrick objects", {
   dimension = 10, multi_layer_file = TRUE)
   stack = raster::brick(stack)
 
-  expect_class(as_task_unsupervised(stack), "Task")
+  expect_class(as_task_unsupervised(stack), "TaskUnsupervised")
 })
 
-test_that("as_task works on RasterStack objects", {
+test_that("as_task_unsupervised works on RasterStack objects", {
   skip_if_not_installed("raster")
   requireNamespace("raster", quietly = TRUE)
 
@@ -44,5 +44,12 @@ test_that("as_task works on RasterStack objects", {
   stack = invoke(raster::stack, x = stack, .opts = allow_partial_matching)
   raster::crs(stack) = "EPSG:4326"
 
-  expect_class(as_task_unsupervised(stack), "Task")
+  expect_class(as_task_unsupervised(stack), "TaskUnsupervised")
+})
+
+test_that("as_task_unsupervised works on sf objects", {
+  vector = sf::read_sf(system.file("extdata", "leipzig_points.gpkg", package = "mlr3spatial"), stringsAsFactors = TRUE)
+  vector$land_cover = NULL
+
+  expect_class(as_task_unsupervised(vector), "TaskUnsupervised")
 })