+++ NEWS.md
@@ -1,5 +1,7 @@
 # mlr3spatial 0.2.1.9000
 
+* feat: add prediction on vector data to `spatial_predict()`.
+
 # mlr3spatial 0.2.1
 
 * fix: add `"space"` and `"time"` column role from mlr3spatiotempcv